requirements.txt
customtkinter
mido
python-rtmidi
soundpad_control
